+++ routes/userStatus.js
@@ -5,6 +5,7 @@ const {
   getAllUserStatus,
   updateUserStatus,
   updateAllUserStatus,
+  getUsersWithOnboardingState,
 } = require("../controllers/userStatus");
 const router = express.Router();
 const authenticate = require("../middlewares/authenticate");
@@ -14,6 +15,7 @@ const { validateUserStatus } = require("../middlewares/validators/userStatus");
 
 router.get("/", getAllUserStatus);
 router.get("/self", authenticate, getUserStatus);
+router.get("/onboarding", authenticate, authorizeRoles([SUPERUSER]), getUsersWithOnboardingState);
 router.get("/:userId", getUserStatus);
 router.patch("/self", authenticate, validateUserStatus, updateUserStatus);
 router.patch("/update", authenticate, authorizeRoles([SUPERUSER]), updateAllUserStatus);
